[259848]  Provide asynchronous progress monitoring https://bugs.eclipse.org/bugs/show_bug.cgi?id=259848

+++ plugins/org.eclipse.net4j/src/org/eclipse/net4j/signal/MonitorProgressIndication.java
@@ -31,8 +31,8 @@ class MonitorProgressIndication extends Indication
   protected void indicating(ExtendedDataInputStream in) throws Exception
   {
     int correlationID = in.readInt();
-    int totalWork = in.readInt();
-    int work = in.readInt();
+    double totalWork = in.readDouble();
+    double work = in.readDouble();
     if (TRACER.isEnabled())
     {
       TRACER.format("Progress of signal {0}: totalWork={1}, work={2}", correlationID, totalWork, work);
